+++ relay-event-schema/src/protocol/span/compat.rs
@@ -75,6 +75,12 @@ impl TryFrom<SpanV2> for CompatSpan {
                 // ignoring meta here is OK, internal attribute set by Relay.
                 compat_span.performance_issues_spans = Annotated::new(*b);
             }
+
+            if let Some(Value::Bool(b)) =
+                attributes.get_value("sentry._internal.performance_issues_spans")
+            {
+                compat_span.performance_issues_spans = Annotated::new(*b);
+            }
         }
 
         // FIXME: preserve _meta everywhere (see a test failure).