+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - No longer writes Spans as trace items. ([#5152](https://github.com/getsentry/relay/pull/5152))
 
+**Features**:
+
+- Add Thread Pool Info context to event schema. ([#5153](https://github.com/getsentry/relay/pull/5153))
+
 ## 25.9.0
 
 **Breaking Changes**:
@@ -19,7 +23,6 @@
 - Add dynamic PII derivation to `metastructure`. ([#5107](https://github.com/getsentry/relay/pull/5107))
 - Detect PII status of attributes based on `sentry-conventions`. ([#5113](https://github.com/getsentry/relay/pull/5113))
 - Add support for an OTLP `/v1/logs` endpoint. This endpoint is gated behind the `organizations:relay-otel-logs-endpoint` feature flag. ([#5130](https://github.com/getsentry/relay/pull/5130))
-- Add Thread Pool Info context to event schema. ([#5153](https://github.com/getsentry/relay/pull/5153))
 
 **Internal**: