+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+**Internal**:
+
+- No longer writes Spans as trace items. ([#5152](https://github.com/getsentry/relay/pull/5152))
+
+**Features**:
+
+- Add Thread Pool Info context to event schema. ([#5153](https://github.com/getsentry/relay/pull/5153))
+
+## 25.9.0
+
 **Breaking Changes**:
 
 - Removes support for the deprecated and early alpha only otel log item type. ([#5082](https://github.com/getsentry/relay/pull/5082))